Added capture timestamp to OnFrameReady event
Added the timestamp of the capture to the OnFrameReady event and store it on the YUV Converter so that it can be accessed via the capture session.

Capture timestamp is needed for any camera use cases that require an accurate camera pose as the capture texture is received ~50ms after it's taken. Using the current camera pose instead of the camera pose when the frame was captured results in instability in use cases like marker tracking or compositing onto the camera texture.